refactor: restructure TechStack section by moving to separate files and improving component organization
