+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import Career from "@/components/sections/career"
 import Contact from "@/components/sections/contact"
 import Introduction from "@/components/sections/introduction"
-import TechStack from "@/components/sections/TechStack"
+import TechStack from "@/components/sections/tech-stack"
 import { Container, Separator } from "@chakra-ui/react"
 
 export const dynamic = 'force-static'
