+++ src/components/sections/tech-stack/index.tsx
@@ -1,47 +1,7 @@
-'use client'
-
 import Paragraph from "@/components/ui/Paragraph"
-import { useColorModeValue } from "@/components/ui/color-mode"
 import { techStacks } from "@/constants"
-import { Box, Flex, Grid, Heading, IconButton, Stack, Tabs, Text } from "@chakra-ui/react"
-import type { ReactElement } from "react"
-
-interface IStackCard {
-  stack: { name: string, icon: ReactElement }
-}
-const StackCard = ({ stack }: IStackCard) => {
-  return (
-    <Box
-      w='full'
-      padding={4}
-      borderColor={useColorModeValue('gray.300', 'gray.700')}
-      borderRadius={5}
-      borderWidth='1px'
-      transition='border-color 0.5s, transform 0.2s cubic-bezier(0.4, 0, 0.2, 1)'
-      cursor='pointer'
-      display='flex'
-      role='group'
-      _hover={{
-        borderColor: 'gray.500',
-        transform: 'translateY(-5px)',
-      }}
-    >
-      <Flex alignItems='center' justifyContent='space-between'>
-        <IconButton
-          aria-label={stack.name}
-          marginRight={3}
-        >
-          {stack.icon}
-        </IconButton>
-        <Flex>
-          <Text textStyle='sm' _hover={{ color: 'gray.500' }}>
-            {stack.name}
-          </Text>
-        </Flex>
-      </Flex>
-    </Box>
-  )
-}
+import { Grid, Heading, Stack, Tabs } from "@chakra-ui/react"
+import StackCard from "./stack-card"
 
 export default function TechStack() {
   return (
